pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nichifier-platform"
version = "0.1.0"
description = "Business intelligence newsletter and report platform"
requires-python = ">=3.11"
authors = [{name = "Nichifier Platform"}]
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "jinja2>=3.1.3",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.6"
]

[tool.setuptools.packages.find]
where = ["."]
